Refactor rename mixins to models_mixins
